added refactor code to course and profile
