+++ ForoHub/src/main/java/com/jsrdev/ForoHub/infrastructure/rest/dto/DeleteResponse.java
@@ -1,4 +1,4 @@
-package com.jsrdev.ForoHub.infrastructure.rest.dto.course;
+package com.jsrdev.ForoHub.infrastructure.rest.dto;
 
 public record DeleteResponse(boolean isDeleted, String message) {
 }
